More method bindings + notes regarding from/to string + partial chng
Note the from/to string implementation is kind of broken. Focus is
on avoiding multiple copying of strings or being given a `malloc`
ptr that then needs to be dealt with.

It's not the biggest issue. I can memcopy to a zig allocation and
then `free` said string. But it should be possible to do better. It
has given me the opportunity to investigate in-depth regard direct
bit twiddling/flipping action.